pyproject.toml
[project]
name = "code-voyagers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.8.0",
    "google-generativeai>=0.8.5",
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "pyprojroot>=0.3.0",
]
